// packages/destination-actions/src/destinations/ripe/alias/generated-types.ts
// Generated file. DO NOT MODIFY IT BY HAND.

export interface Payload {
  /**
   * The anonymized user id
   */
  anonymousId?: string | null
  /**
   * The ID associated with the user
   */
  userId: string
  /**
   * The alias of the user
   */
  alias: string
  /**
   * The timestamp of the event
   */
  timestamp?: string | number
}